Standings Table "component"
Moved the standings function into its own file so it works sort of like
a component, even though it's just a JS functinon that returns html.
